email_assistant/pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "langchain>=1.0.7",
    "langchain-google-genai>=3.0.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
